moves disconnect btn to header
