+++ packages/wallet-ui/src/components/ui/SideBar/SideBar.view.tsx
@@ -1,11 +1,5 @@
-import { useCallback, useRef, useState } from 'react';
-import { useAppDispatch, useAppSelector } from '../../../hooks/redux';
-import { resetNetwork } from '../../../slices/networkSlice';
-import {
-  resetWallet,
-  setForceReconnect,
-  setWalletConnection,
-} from '../../../slices/walletSlice';
+import { useRef, useState } from 'react';
+import { useAppSelector } from '../../../hooks/redux';
 import { AccountAddress } from '../AccountAddress';
 import { AccountDetailsModal } from '../AccountDetailsModal';
 import { AssetsList } from '../AssetsList';
@@ -29,16 +23,8 @@ type Props = {
 export const SideBarView = ({ address }: Props) => {
   const [accountDetailsOpen, setAccountDetailsOpen] = useState(false);
   const wallet = useAppSelector((state) => state.wallet);
-  const dispatch = useAppDispatch();
 
   const ref = useRef<HTMLDivElement>();
-
-  const disconnect = useCallback(() => {
-    dispatch(setWalletConnection(false));
-    dispatch(setForceReconnect(true));
-    dispatch(resetWallet());
-    dispatch(resetNetwork());
-  }, []);
 
   return (
     <Wrapper>
@@ -60,13 +46,6 @@ export const SideBarView = ({ address }: Props) => {
               onClick={() => setAccountDetailsOpen(true)}
             >
               Account details
-            </AccountDetailButton>
-            <AccountDetailButton
-              backgroundTransparent
-              iconLeft="right-to-bracket"
-              onClick={() => disconnect()}
-            >
-              Disconnect
             </AccountDetailButton>
           </AccountDetailsContent>
         }
